LayoutTests/imported/w3c/web-platform-tests/svg/types/scripted/SVGAnimatedLength-initial-values-expected.txt

CONSOLE MESSAGE: Error: Invalid value for <circle> attribute cx="foobar"
CONSOLE MESSAGE: Error: Invalid value for <circle> attribute cy="foobar"
CONSOLE MESSAGE: Error: Invalid value for <circle> attribute r="foobar"
CONSOLE MESSAGE: Error: Invalid value for <ellipse> attribute cx="foobar"
CONSOLE MESSAGE: Error: Invalid value for <ellipse> attribute cy="foobar"
CONSOLE MESSAGE: Error: Invalid value for <ellipse> attribute rx="foobar"
CONSOLE MESSAGE: Error: Invalid value for <ellipse> attribute ry="foobar"
CONSOLE MESSAGE: Error: Invalid value for <filter> attribute x="foobar"
CONSOLE MESSAGE: Error: Invalid value for <filter> attribute y="foobar"
CONSOLE MESSAGE: Error: Invalid value for <filter> attribute width="foobar"
CONSOLE MESSAGE: Error: Invalid value for <filter> attribute height="foobar"
CONSOLE MESSAGE: Error: Invalid value for <feBlend> attribute x="foobar"
CONSOLE MESSAGE: Error: Invalid value for <feBlend> attribute y="foobar"
CONSOLE MESSAGE: Error: Invalid value for <feBlend> attribute width="foobar"
CONSOLE MESSAGE: Error: Invalid value for <feBlend> attribute height="foobar"
CONSOLE MESSAGE: Error: Invalid value for <foreignObject> attribute x="foobar"
CONSOLE MESSAGE: Error: Invalid value for <foreignObject> attribute y="foobar"
CONSOLE MESSAGE: Error: Invalid value for <foreignObject> attribute width="foobar"
CONSOLE MESSAGE: Error: Invalid value for <foreignObject> attribute height="foobar"
CONSOLE MESSAGE: Error: Invalid value for <image> attribute x="foobar"
CONSOLE MESSAGE: Error: Invalid value for <image> attribute y="foobar"
CONSOLE MESSAGE: Error: Invalid value for <image> attribute width="foobar"
CONSOLE MESSAGE: Error: Invalid value for <image> attribute height="foobar"
CONSOLE MESSAGE: Error: Invalid value for <line> attribute x1="foobar"
CONSOLE MESSAGE: Error: Invalid value for <line> attribute y1="foobar"
CONSOLE MESSAGE: Error: Invalid value for <line> attribute x2="foobar"
CONSOLE MESSAGE: Error: Invalid value for <line> attribute y2="foobar"
CONSOLE MESSAGE: Error: Invalid value for <linearGradient> attribute x1="foobar"
CONSOLE MESSAGE: Error: Invalid value for <linearGradient> attribute y1="foobar"
CONSOLE MESSAGE: Error: Invalid value for <linearGradient> attribute x2="foobar"
CONSOLE MESSAGE: Error: Invalid value for <linearGradient> attribute y2="foobar"
CONSOLE MESSAGE: Error: Invalid value for <marker> attribute refX="foobar"
CONSOLE MESSAGE: Error: Invalid value for <marker> attribute refY="foobar"
CONSOLE MESSAGE: Error: Invalid value for <marker> attribute markerWidth="foobar"
CONSOLE MESSAGE: Error: Invalid value for <marker> attribute markerHeight="foobar"
CONSOLE MESSAGE: Error: Invalid value for <mask> attribute x="foobar"
CONSOLE MESSAGE: Error: Invalid value for <mask> attribute y="foobar"
CONSOLE MESSAGE: Error: Invalid value for <mask> attribute width="foobar"
CONSOLE MESSAGE: Error: Invalid value for <mask> attribute height="foobar"
CONSOLE MESSAGE: Error: Invalid value for <pattern> attribute x="foobar"
CONSOLE MESSAGE: Error: Invalid value for <pattern> attribute y="foobar"
CONSOLE MESSAGE: Error: Invalid value for <pattern> attribute width="foobar"
CONSOLE MESSAGE: Error: Invalid value for <pattern> attribute height="foobar"
CONSOLE MESSAGE: Error: Invalid value for <radialGradient> attribute cx="foobar"
CONSOLE MESSAGE: Error: Invalid value for <radialGradient> attribute cy="foobar"
CONSOLE MESSAGE: Error: Invalid value for <radialGradient> attribute r="foobar"
CONSOLE MESSAGE: Error: Invalid value for <radialGradient> attribute fr="foobar"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute x="foobar"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute y="foobar"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute width="foobar"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="foobar"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute rx="foobar"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute ry="foobar"
CONSOLE MESSAGE: Error: Invalid value for <svg> attribute x="foobar"
CONSOLE MESSAGE: Error: Invalid value for <svg> attribute y="foobar"
CONSOLE MESSAGE: Error: Invalid value for <svg> attribute width="foobar"
CONSOLE MESSAGE: Error: Invalid value for <svg> attribute height="foobar"
CONSOLE MESSAGE: Error: Invalid value for <text> attribute textLength="foobar"
CONSOLE MESSAGE: Error: Invalid value for <textPath> attribute startOffset="foobar"
CONSOLE MESSAGE: Error: Invalid value for <use> attribute x="foobar"
CONSOLE MESSAGE: Error: Invalid value for <use> attribute y="foobar"
CONSOLE MESSAGE: Error: Invalid value for <use> attribute width="foobar"
CONSOLE MESSAGE: Error: Invalid value for <use> attribute height="foobar"

PASS SVGAnimatedLength, initial values, SVGCircleElement.prototype.cx (remove)
PASS SVGAnimatedLength, initial values, SVGCircleElement.prototype.cx (invalid value)
PASS SVGAnimatedLength, initial values, SVGCircleElement.prototype.cy (remove)
PASS SVGAnimatedLength, initial values, SVGCircleElement.prototype.cy (invalid value)
PASS SVGAnimatedLength, initial values, SVGCircleElement.prototype.r (remove)
PASS SVGAnimatedLength, initial values, SVGCircleElement.prototype.r (invalid value)
PASS SVGAnimatedLength, initial values, SVGEllipseElement.prototype.cx (remove)
PASS SVGAnimatedLength, initial values, SVGEllipseElement.prototype.cx (invalid value)
PASS SVGAnimatedLength, initial values, SVGEllipseElement.prototype.cy (remove)
PASS SVGAnimatedLength, initial values, SVGEllipseElement.prototype.cy (invalid value)
PASS SVGAnimatedLength, initial values, SVGEllipseElement.prototype.rx (remove)
PASS SVGAnimatedLength, initial values, SVGEllipseElement.prototype.rx (invalid value)
PASS SVGAnimatedLength, initial values, SVGEllipseElement.prototype.ry (remove)
PASS SVGAnimatedLength, initial values, SVGEllipseElement.prototype.ry (invalid value)
FAIL SVGAnimatedLength, initial values, SVGFilterElement.prototype.x (remove) assert_equals: initial after expected "-10%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterElement.prototype.x (invalid value) assert_equals: initial after expected "-10%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterElement.prototype.y (remove) assert_equals: initial after expected "-10%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterElement.prototype.y (invalid value) assert_equals: initial after expected "-10%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterElement.prototype.width (remove) assert_equals: initial after expected "120%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterElement.prototype.width (invalid value) assert_equals: initial after expected "120%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterElement.prototype.height (remove) assert_equals: initial after expected "120%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterElement.prototype.height (invalid value) assert_equals: initial after expected "120%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterPrimitiveStandardAttributes.prototype.x (remove) assert_equals: initial after expected "0%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterPrimitiveStandardAttributes.prototype.x (invalid value) assert_equals: initial after expected "0%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterPrimitiveStandardAttributes.prototype.y (remove) assert_equals: initial after expected "0%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterPrimitiveStandardAttributes.prototype.y (invalid value) assert_equals: initial after expected "0%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterPrimitiveStandardAttributes.prototype.width (remove) assert_equals: initial after expected "100%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterPrimitiveStandardAttributes.prototype.width (invalid value) assert_equals: initial after expected "100%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterPrimitiveStandardAttributes.prototype.height (remove) assert_equals: initial after expected "100%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGFilterPrimitiveStandardAttributes.prototype.height (invalid value) assert_equals: initial after expected "100%" but got "0"
PASS SVGAnimatedLength, initial values, SVGForeignObjectElement.prototype.x (remove)
PASS SVGAnimatedLength, initial values, SVGForeignObjectElement.prototype.x (invalid value)
PASS SVGAnimatedLength, initial values, SVGForeignObjectElement.prototype.y (remove)
PASS SVGAnimatedLength, initial values, SVGForeignObjectElement.prototype.y (invalid value)
PASS SVGAnimatedLength, initial values, SVGForeignObjectElement.prototype.width (remove)
PASS SVGAnimatedLength, initial values, SVGForeignObjectElement.prototype.width (invalid value)
PASS SVGAnimatedLength, initial values, SVGForeignObjectElement.prototype.height (remove)
PASS SVGAnimatedLength, initial values, SVGForeignObjectElement.prototype.height (invalid value)
PASS SVGAnimatedLength, initial values, SVGImageElement.prototype.x (remove)
PASS SVGAnimatedLength, initial values, SVGImageElement.prototype.x (invalid value)
PASS SVGAnimatedLength, initial values, SVGImageElement.prototype.y (remove)
PASS SVGAnimatedLength, initial values, SVGImageElement.prototype.y (invalid value)
PASS SVGAnimatedLength, initial values, SVGImageElement.prototype.width (remove)
PASS SVGAnimatedLength, initial values, SVGImageElement.prototype.width (invalid value)
PASS SVGAnimatedLength, initial values, SVGImageElement.prototype.height (remove)
PASS SVGAnimatedLength, initial values, SVGImageElement.prototype.height (invalid value)
PASS SVGAnimatedLength, initial values, SVGLineElement.prototype.x1 (remove)
PASS SVGAnimatedLength, initial values, SVGLineElement.prototype.x1 (invalid value)
PASS SVGAnimatedLength, initial values, SVGLineElement.prototype.y1 (remove)
PASS SVGAnimatedLength, initial values, SVGLineElement.prototype.y1 (invalid value)
PASS SVGAnimatedLength, initial values, SVGLineElement.prototype.x2 (remove)
PASS SVGAnimatedLength, initial values, SVGLineElement.prototype.x2 (invalid value)
PASS SVGAnimatedLength, initial values, SVGLineElement.prototype.y2 (remove)
PASS SVGAnimatedLength, initial values, SVGLineElement.prototype.y2 (invalid value)
FAIL SVGAnimatedLength, initial values, SVGLinearGradientElement.prototype.x1 (remove) assert_equals: initial after expected "0%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGLinearGradientElement.prototype.x1 (invalid value) assert_equals: initial after expected "0%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGLinearGradientElement.prototype.y1 (remove) assert_equals: initial after expected "0%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGLinearGradientElement.prototype.y1 (invalid value) assert_equals: initial after expected "0%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGLinearGradientElement.prototype.x2 (remove) assert_equals: initial after expected "100%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGLinearGradientElement.prototype.x2 (invalid value) assert_equals: initial after expected "100%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGLinearGradientElement.prototype.y2 (remove) assert_equals: initial after expected "0%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGLinearGradientElement.prototype.y2 (invalid value) assert_equals: initial after expected "0%" but got "0"
PASS SVGAnimatedLength, initial values, SVGMarkerElement.prototype.refX (remove)
PASS SVGAnimatedLength, initial values, SVGMarkerElement.prototype.refX (invalid value)
PASS SVGAnimatedLength, initial values, SVGMarkerElement.prototype.refY (remove)
PASS SVGAnimatedLength, initial values, SVGMarkerElement.prototype.refY (invalid value)
FAIL SVGAnimatedLength, initial values, SVGMarkerElement.prototype.markerWidth (remove) assert_equals: initial after expected "3" but got "0"
FAIL SVGAnimatedLength, initial values, SVGMarkerElement.prototype.markerWidth (invalid value) assert_equals: initial after expected "3" but got "0"
FAIL SVGAnimatedLength, initial values, SVGMarkerElement.prototype.markerHeight (remove) assert_equals: initial after expected "3" but got "0"
FAIL SVGAnimatedLength, initial values, SVGMarkerElement.prototype.markerHeight (invalid value) assert_equals: initial after expected "3" but got "0"
FAIL SVGAnimatedLength, initial values, SVGMaskElement.prototype.x (remove) assert_equals: initial after expected "-10%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGMaskElement.prototype.x (invalid value) assert_equals: initial after expected "-10%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGMaskElement.prototype.y (remove) assert_equals: initial after expected "-10%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGMaskElement.prototype.y (invalid value) assert_equals: initial after expected "-10%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGMaskElement.prototype.width (remove) assert_equals: initial after expected "120%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGMaskElement.prototype.width (invalid value) assert_equals: initial after expected "120%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGMaskElement.prototype.height (remove) assert_equals: initial after expected "120%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGMaskElement.prototype.height (invalid value) assert_equals: initial after expected "120%" but got "0"
PASS SVGAnimatedLength, initial values, SVGPatternElement.prototype.x (remove)
PASS SVGAnimatedLength, initial values, SVGPatternElement.prototype.x (invalid value)
PASS SVGAnimatedLength, initial values, SVGPatternElement.prototype.y (remove)
PASS SVGAnimatedLength, initial values, SVGPatternElement.prototype.y (invalid value)
PASS SVGAnimatedLength, initial values, SVGPatternElement.prototype.width (remove)
PASS SVGAnimatedLength, initial values, SVGPatternElement.prototype.width (invalid value)
PASS SVGAnimatedLength, initial values, SVGPatternElement.prototype.height (remove)
PASS SVGAnimatedLength, initial values, SVGPatternElement.prototype.height (invalid value)
FAIL SVGAnimatedLength, initial values, SVGRadialGradientElement.prototype.cx (remove) assert_equals: initial after expected "50%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGRadialGradientElement.prototype.cx (invalid value) assert_equals: initial after expected "50%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGRadialGradientElement.prototype.cy (remove) assert_equals: initial after expected "50%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGRadialGradientElement.prototype.cy (invalid value) assert_equals: initial after expected "50%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGRadialGradientElement.prototype.r (remove) assert_equals: initial after expected "50%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGRadialGradientElement.prototype.r (invalid value) assert_equals: initial after expected "50%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGRadialGradientElement.prototype.fx (remove) assert_equals: initial before expected "50%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGRadialGradientElement.prototype.fx (invalid value) assert_equals: initial before expected "50%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGRadialGradientElement.prototype.fy (remove) assert_equals: initial before expected "50%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGRadialGradientElement.prototype.fy (invalid value) assert_equals: initial before expected "50%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGRadialGradientElement.prototype.fr (remove) assert_equals: initial after expected "0%" but got "0"
FAIL SVGAnimatedLength, initial values, SVGRadialGradientElement.prototype.fr (invalid value) assert_equals: initial after expected "0%" but got "0"
PASS SVGAnimatedLength, initial values, SVGRectElement.prototype.x (remove)
PASS SVGAnimatedLength, initial values, SVGRectElement.prototype.x (invalid value)
PASS SVGAnimatedLength, initial values, SVGRectElement.prototype.y (remove)
PASS SVGAnimatedLength, initial values, SVGRectElement.prototype.y (invalid value)
PASS SVGAnimatedLength, initial values, SVGRectElement.prototype.width (remove)
PASS SVGAnimatedLength, initial values, SVGRectElement.prototype.width (invalid value)
PASS SVGAnimatedLength, initial values, SVGRectElement.prototype.height (remove)
PASS SVGAnimatedLength, initial values, SVGRectElement.prototype.height (invalid value)
PASS SVGAnimatedLength, initial values, SVGRectElement.prototype.rx (remove)
PASS SVGAnimatedLength, initial values, SVGRectElement.prototype.rx (invalid value)
PASS SVGAnimatedLength, initial values, SVGRectElement.prototype.ry (remove)
PASS SVGAnimatedLength, initial values, SVGRectElement.prototype.ry (invalid value)
PASS SVGAnimatedLength, initial values, SVGSVGElement.prototype.x (remove)
PASS SVGAnimatedLength, initial values, SVGSVGElement.prototype.x (invalid value)
PASS SVGAnimatedLength, initial values, SVGSVGElement.prototype.y (remove)
PASS SVGAnimatedLength, initial values, SVGSVGElement.prototype.y (invalid value)
PASS SVGAnimatedLength, initial values, SVGSVGElement.prototype.width (remove)
PASS SVGAnimatedLength, initial values, SVGSVGElement.prototype.width (invalid value)
PASS SVGAnimatedLength, initial values, SVGSVGElement.prototype.height (remove)
PASS SVGAnimatedLength, initial values, SVGSVGElement.prototype.height (invalid value)
PASS SVGAnimatedLength, initial values, SVGTextContentElement.prototype.textLength (remove)
PASS SVGAnimatedLength, initial values, SVGTextContentElement.prototype.textLength (invalid value)
PASS SVGAnimatedLength, initial values, SVGTextPathElement.prototype.startOffset (remove)
PASS SVGAnimatedLength, initial values, SVGTextPathElement.prototype.startOffset (invalid value)
PASS SVGAnimatedLength, initial values, SVGUseElement.prototype.x (remove)
PASS SVGAnimatedLength, initial values, SVGUseElement.prototype.x (invalid value)
PASS SVGAnimatedLength, initial values, SVGUseElement.prototype.y (remove)
PASS SVGAnimatedLength, initial values, SVGUseElement.prototype.y (invalid value)
PASS SVGAnimatedLength, initial values, SVGUseElement.prototype.width (remove)
PASS SVGAnimatedLength, initial values, SVGUseElement.prototype.width (invalid value)
PASS SVGAnimatedLength, initial values, SVGUseElement.prototype.height (remove)
PASS SVGAnimatedLength, initial values, SVGUseElement.prototype.height (invalid value)
